crt-royale: Add .cgp presets that work with Intel HD 4000 Graphics out of the box. Move source files into the src directory, so users can more easily find the user-settings.h file (etc.).  Update README.TXT accordingly. Finish replacing levels_underestimate with bloom_underestimate_levels (missed a spot earlier).
