+++ crt/crt-royale/src/crt-royale-bloom-approx-fake-bloom.cg
