+++ crt/crt-royale/src/crt-royale-bloom-approx-intel.cg
@@ -1,0 +1,2 @@
+#define INTEGRATED_GRAPHICS_COMPATIBILITY_MODE
+#include "crt-royale-bloom-approx.cg"
