+++ crt/crt-royale/src/crt-royale-bloom-approx.cg
@@ -19,11 +19,11 @@
 
 //////////////////////////////////  INCLUDES  //////////////////////////////////
 
-#include "user-settings.h"
+#include "../user-settings.h"
 #include "derived-settings-and-constants.h"
 #include "bind-shader-params.h"
-#include "../../include/gamma-management.h"
-#include "../../include/blur-functions.h"
+#include "../../../include/gamma-management.h"
+#include "../../../include/blur-functions.h"
 #include "scanline-functions.h"
 #include "bloom-functions.h"
