+++ crt/crt-royale/src/crt-royale-bloom-vertical.cg
@@ -19,14 +19,14 @@
 
 /////////////////////////////  SETTINGS MANAGEMENT  ////////////////////////////
 
-#include "user-settings.h"
+#include "../user-settings.h"
 #include "derived-settings-and-constants.h"
 #include "bind-shader-params.h"
 
 
 //////////////////////////////////  INCLUDES  //////////////////////////////////
 
-#include "../../include/gamma-management.h"
+#include "../../../include/gamma-management.h"
 #include "bloom-functions.h"
 #include "phosphor-mask-resizing.h"
