+++ crt/crt-royale/src/crt-royale-brightpass.cg
@@ -19,15 +19,15 @@
 
 /////////////////////////////  SETTINGS MANAGEMENT  ////////////////////////////
 
-#include "user-settings.h"
+#include "../user-settings.h"
 #include "derived-settings-and-constants.h"
 #include "bind-shader-params.h"
 
 
 //////////////////////////////////  INCLUDES  //////////////////////////////////
 
-#include "../../include/gamma-management.h"
-#include "../../include/blur-functions.h"
+#include "../../../include/gamma-management.h"
+#include "../../../include/blur-functions.h"
 #include "phosphor-mask-resizing.h"
 #include "scanline-functions.h"
 #include "bloom-functions.h"
