+++ crt/crt-royale/src/crt-royale-geometry-aa-last-pass-intel.cg
@@ -1,0 +1,2 @@
+#define INTEGRATED_GRAPHICS_COMPATIBILITY_MODE
+#include "crt-royale-geometry-aa-last-pass.cg"
