+++ crt/crt-royale/src/crt-royale-mask-resize-vertical.cg
@@ -19,7 +19,7 @@
 
 /////////////////////////////  SETTINGS MANAGEMENT  ////////////////////////////
 
-#include "user-settings.h"
+#include "../user-settings.h"
 #include "derived-settings-and-constants.h"
 #include "bind-shader-params.h"
