+++ crt/crt-royale/src/crt-royale-scanlines-horizontal-apply-mask-intel.cg
@@ -1,0 +1,2 @@
+#define INTEGRATED_GRAPHICS_COMPATIBILITY_MODE
+#include "crt-royale-scanlines-horizontal-apply-mask.cg"
