+++ crt/crt-royale/src/crt-royale-scanlines-vertical-interlacing.cg
@@ -19,11 +19,11 @@
 
 //////////////////////////////////  INCLUDES  //////////////////////////////////
 
-#include "user-settings.h"
+#include "../user-settings.h"
 #include "derived-settings-and-constants.h"
 #include "bind-shader-params.h"
 #include "scanline-functions.h"
-#include "../../include/gamma-management.h"
+#include "../../../include/gamma-management.h"
 
 
 /////////////////////////////////  STRUCTURES  /////////////////////////////////
