+++ crt/crt-royale/src/phosphor-mask-resizing.h
@@ -22,7 +22,7 @@
 
 //////////////////////////////////  INCLUDES  //////////////////////////////////
 
-#include "user-settings.h"
+#include "../user-settings.h"
 #include "derived-settings-and-constants.h"
 
 /////////////////////////////  CODEPATH SELECTION  /////////////////////////////
