+++ crt/crt-royale/src/scanline-functions.h
@@ -22,10 +22,10 @@
 
 //////////////////////////////////  INCLUDES  //////////////////////////////////
 
-#include "user-settings.h"
+#include "../user-settings.h"
 #include "derived-settings-and-constants.h"
-#include "../../include/special-functions.h"
-#include "../../include/gamma-management.h"
+#include "../../../include/special-functions.h"
+#include "../../../include/gamma-management.h"
 
 
 /////////////////////////////  SCANLINE FUNCTIONS  /////////////////////////////
