+++ crt/crt-royale/src/tex2Dantialias.h
@@ -203,7 +203,7 @@
 
 //////////////////////////////////  INCLUDES  //////////////////////////////////
 
-#include "../../include/gamma-management.h"
+#include "../../../include/gamma-management.h"
 
 
 //////////////////////////////////  CONSTANTS  /////////////////////////////////
