+++ crt/crt-royale/src/user-cgp-constants.h
